Move NetworkConnectionBase
